[Frontend] Relationship Number widget refetch (#14448)
