+++ opencti-platform/opencti-front/src/components/dashboard/DashboardViz.tsx
@@ -41,6 +41,7 @@ const DashboardViz = ({
             config={config}
             popover={popover}
             host={host}
+            refreshRate={refreshRate}
           />
         )}
         {perspective === 'audits' && (
