--- opencti-platform/opencti-front/src/private/components/common/stix_relationships/StixRelationshipsNumber.jsx
@@ -1,137 +1,0 @@
-import { graphql } from 'react-relay';
-import { QueryRenderer } from '../../../../relay/environment';
-import { useFormatter } from '../../../../components/i18n';
-import { dayAgo } from '../../../../utils/Time';
-import { buildFiltersAndOptionsForWidgets } from '../../../../utils/filters/filtersUtils';
-import WidgetContainer from '../../../../components/dashboard/WidgetContainer';
-import WidgetNoData from '../../../../components/dashboard/WidgetNoData';
-import Loader, { LoaderVariant } from '../../../../components/Loader';
-import useEntityTranslation from '../../../../utils/hooks/useEntityTranslation';
-import WidgetNumber from '../../../../components/dashboard/WidgetNumber';
-import useDashboardViz from '../../../../components/dashboard/useDashboardViz';
-import WidgetNoHostEntity from '../../../../components/dashboard/WidgetNoHostEntity';
-
-const stixRelationshipsNumberNumberQuery = graphql`
-    query StixRelationshipsNumberNumberSeriesQuery(
-        $dateAttribute: String
-        $noDirection: Boolean
-        $endDate: DateTime
-        $onlyInferred: Boolean
-        $fromOrToId: [String]
-        $elementWithTargetTypes: [String]
-        $fromId: [String]
-        $fromRole: String
-        $fromTypes: [String]
-        $toId: [String]
-        $toRole: String
-        $toTypes: [String]
-        $relationship_type: [String]
-        $confidences: [Int]
-        $search: String
-        $filters: FilterGroup
-        $dynamicFrom: FilterGroup
-        $dynamicTo: FilterGroup
-    ) {
-        stixRelationshipsNumber(
-            dateAttribute: $dateAttribute
-            noDirection: $noDirection
-            endDate: $endDate
-            onlyInferred: $onlyInferred
-            fromOrToId: $fromOrToId
-            elementWithTargetTypes: $elementWithTargetTypes
-            fromId: $fromId
-            fromRole: $fromRole
-            fromTypes: $fromTypes
-            toId: $toId
-            toRole: $toRole
-            toTypes: $toTypes
-            relationship_type: $relationship_type
-            confidences: $confidences
-            search: $search
-            filters: $filters
-            dynamicFrom: $dynamicFrom
-            dynamicTo: $dynamicTo
-        ) {
-            total
-            count
-        }
-    }
-`;
-
-const StixRelationshipsNumber = ({
-  startDate,
-  endDate,
-  dataSelection,
-  parameters = {},
-  entityType,
-  popover,
-  variant,
-  height,
-  host,
-}) => {
-  const { t_i18n } = useFormatter();
-  const { translateEntityType } = useEntityTranslation();
-
-  const title = parameters.title ?? t_i18n('Entities number');
-  const translatedTitle = translateEntityType(title);
-
-  const { resolvedDataSelection, isMissingHostEntity, isPreviewMode } = useDashboardViz({
-    perspective: 'relationships',
-    dataSelection,
-    host,
-  });
-
-  const selection = resolvedDataSelection[0];
-  const dateAttribute = selection.date_attribute && selection.date_attribute.length > 0
-    ? selection.date_attribute
-    : 'created_at';
-  const { filters } = buildFiltersAndOptionsForWidgets(selection.filters, { startDate, endDate, dateAttribute, isKnowledgeRelationshipWidget: true });
-
-  return (
-    <WidgetContainer
-      padding="medium"
-      height={height}
-      title={t_i18n('Relationships number')}
-      variant={variant}
-      action={popover}
-      showPreviewTag={isPreviewMode}
-    >
-      {
-        isMissingHostEntity
-          ? <WidgetNoHostEntity host={host} />
-          : (
-              <QueryRenderer
-                query={stixRelationshipsNumberNumberQuery}
-                variables={{
-                  filters,
-                  startDate,
-                  dateAttribute,
-                  endDate: dayAgo(),
-                  dynamicFrom: selection.dynamicFrom,
-                  dynamicTo: selection.dynamicTo,
-                }}
-                render={({ props }) => {
-                  if (props && props.stixRelationshipsNumber) {
-                    const { total, count } = props.stixRelationshipsNumber;
-                    return (
-                      <WidgetNumber
-                        entityType={entityType}
-                        label={translatedTitle}
-                        value={total}
-                        diffLabel={t_i18n('24 hours')}
-                        diffValue={total - count}
-                      />
-                    );
-                  }
-                  if (props) {
-                    return <WidgetNoData />;
-                  }
-                  return <Loader variant={LoaderVariant.inElement} />;
-                }}
-              />
-            )}
-    </WidgetContainer>
-  );
-};
-
-export default StixRelationshipsNumber;
